backend/requirements.txt
# FastAPI y dependencias web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de datos
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Machine Learning y AI
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
joblib==1.3.2

# NLP y procesamiento de texto
nltk==3.8.1
textblob==0.17.1

# Utilidades
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0

# CORS
python-cors==1.0.1

# Desarrollo y testing
pytest==7.4.3
pytest-asyncio==0.21.1